+++ src/maths/fft/matlib.h
@@ -1,6 +1,6 @@
 /* a few routines from a vector/matrix library */
 
-void xpose(float *indata, long iRsiz, float *outdata, long oRsiz, long Nrows, long Ncols);
+void xpose(double *indata, long iRsiz, double *outdata, long oRsiz, long Nrows, long Ncols);
 /* not in-place matrix transpose	*/
 /* INPUTS */
 /* *indata = input data array	*/
@@ -11,7 +11,7 @@ void xpose(float *indata, long iRsiz, float *outdata, long oRsiz, long Nrows, lo
 /* OUTPUTS */
 /* *outdata = output data array	*/
 
-void cxpose(float *indata, long iRsiz, float *outdata, long oRsiz, long Nrows, long Ncols);
+void cxpose(double *indata, long iRsiz, double *outdata, long oRsiz, long Nrows, long Ncols);
 /* not in-place complex matrix transpose	*/
 /* INPUTS */
 /* *indata = input data array	*/
@@ -22,7 +22,7 @@ void cxpose(float *indata, long iRsiz, float *outdata, long oRsiz, long Nrows, l
 /* OUTPUTS */
 /* *outdata = output data array	*/
 
-void cvprod(float *a, float *b, float *out, long N);
+void cvprod(double *a, double *b, double *out, long N);
 /* complex vector product, can be in-place */
 /* product of complex vector *a times complex vector *b */
 /* INPUTS */
